% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA-demo_chemicals.R
\docType{data}
\name{demo_chemicals}
\alias{demo_chemicals}
\title{Demo chemicals}
\format{A 5-by-2 \code{data.frame}.}
\usage{
data(demo_chemicals)
}
\description{
Small data set of five chemicals and their InChIKey for demonstration purposes.
}
\details{
These five chemicals were chosen randomly. The verified InChIKeys were extracted from Wikipedia pages.
}
\examples{
data(demo_chemicals)
demo_chemicals
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
\keyword{datasets}
