% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONALITY-get_apikey.R
\name{get_apikey}
\alias{get_apikey}
\title{Retrieve a ChemSpider API Key}
\usage{
get_apikey(service = "ChemSpider API Key", username = NULL)
}
\arguments{
\item{service}{The name of the key ring service. Defaults to "ChemSpider API Key".}

\item{username}{The username. Defaults to the username as returned by \code{Sys.info()}.}
}
\value{
Securely return the ChemSpider API key as character string.
}
\description{
Functionality to retrieve a ChemSpider API key which has been previously stored.
}
\details{
This function is useful to access your ChemSpider API key.\cr
\cr
To save the ChemSpider API key, the function \code{chemspiderapi::save_apikey()} is provided.
}
\examples{
\dontrun{
## get an API key
get_apikey()
}
}
