% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-get_formula_batch_queryId_results.R
\name{get_formula_batch_queryId_results}
\alias{get_formula_batch_queryId_results}
\title{GET the results of a formula batch query from ChemSpider}
\usage{
get_formula_batch_queryId_results(queryId, status, apikey)
}
\arguments{
\item{queryId}{A valid 36-character query ID, as returned by \code{chemspiderapi::post_formula_batch()}.}

\item{status}{status A character string indicating the query status as returned by \code{chemspiderapi::get_formula_batch_queryId_status()}.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
Returns the (integer) ChemSpider IDs
}
\description{
GET the results of a formula batch query from ChemSpider after \code{chemspiderapi::get_formula_batch_queryId_status()} returns \code{"Complete"}.
}
\details{
Before running \code{chemspiderapi::get_formula_batch_queryId_results()}, make sure \code{chemspiderapi::get_formula_batch_queryId_status()} returns \code{"Complete"}!\cr
\cr
If successful, returns a data frame with the query results; in case the response is a single value, e.g., a ChemSpider ID, it is returned as single vector.\cr
\cr
If not successful, returns \code{NA}.\cr
\cr
This function is fully \code{tidyverse} compatible, e.g., for use in \code{purrr::map_int()}.
}
\examples{
\dontrun{
## Obtain the result from a mass batch query
apikey <- "a valid 32-character ChemSpider apikey"
queryId <- "a valid 36-character ChemSpider queryId"
status <- get_formula_batch_queryId_status(queryId = queryId, apikey = apikey)
get_formula_batch_queryId_results(queryId = queryId, status = status, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/filter/formula/batch/{queryId}/results}
}
