% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-get_mass_batch_queryId_status.R
\name{get_mass_batch_queryId_status}
\alias{get_mass_batch_queryId_status}
\title{GET the status for a mass batch query from ChemSpider}
\usage{
get_mass_batch_queryId_status(queryId, count = TRUE, message = TRUE, apikey)
}
\arguments{
\item{queryId}{A valid 36-character ChemSpider query ID obtained from \code{chemspiderapi::post_mass_batch()}.}

\item{count}{\code{logical}: Should the count of the results be returned (ChemSpider default)?}

\item{message}{\code{logical}: Should the message be returned (ChemSpider default)?}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
Returns a character vector with the status of the query; ideally, "Complete".
}
\description{
This function is used to return the status of a query from \code{chemspiderapi::post_mass_batch()}.
}
\details{
Call this endpoint with a \code{queryId} obtained from \code{chemspiderapi::post_mass_batch()}.\cr
\cr
If the query is still ongoing, returns a warning and a character vector of the query status as \code{Incomplete}. It is recommended to wait at least ten seconds before checking the status again.\cr
\cr
If the query is finalized, returns a data frame of the query status with \code{status}, \code{count} and \code{message}. The \code{status} can be either \code{Complete}, \code{Suspended}, \code{Failed}, or \code{Not Found}.\cr
\cr
Says ChemSpider:\cr
\cr
\emph{"A status of Suspended can be caused if the results could not be compiled within a reasonable amount of time. Create a new filter request with more restrictive parameters.\cr
\cr
A status of Failed can be caused if the backend system could not compile the results. Create a new filter request and, if the same outcome occurs, apply more restrictive parameters.\cr
\cr
A status of Not Found can be caused if the Query ID has not been registered or has expired. Create a new filter request."}\cr
\cr
If both \code{count} and \code{message} are set to \code{FALSE}, \code{chemspiderapi::get_mass_batch_queryId_status()} returns the \code{status} as character vector.\cr
\cr
If the status is \code{Complete}, the results of the query can be obtained from \code{chemspiderapi::get_mass_batch_results()}.\cr
\cr
This function is fully \code{tidyverse} compatible, e.g., for use in \code{purrr::map_chr()}.\cr
}
\examples{
\dontrun{
## Get the status of a mass batch query from ChemSpider
queryId <- "a valid 36-character ChemSpider queryId"
apikey <- "a valid 32-character ChemSpider apikey"
get_mass_batch_queryId_status(queryId = queryId, apikey = apikey)}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/filter/mass/batch/{queryId}/status}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
