% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-get_query_id_results_sdf.R
\name{get_query_id_results_sdf}
\alias{get_query_id_results_sdf}
\title{Get a gzipped SDF file for a ChemSpider query}
\usage{
get_query_id_results_sdf(
  query_id,
  status,
  apikey,
  decode = TRUE,
  simplify = FALSE
)
}
\arguments{
\item{query_id}{A valid 36-character ChemSpider \code{queryId}.}

\item{status}{A character string indicating the query status as returned by \code{get_queryId_status()}.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}

\item{decode}{\code{logical}: should the base64-encoded gzipped file be decoded? Defaults to \code{TRUE}.}

\item{simplify}{\code{logical}: should the results be simplified to a vector? Defaults to \code{FALSE}.}
}
\value{
Returns a (base64-encoded) character vector
}
\description{
This function is used to download a single .sdf file from ChemSpider after \code{get_queryId_status()} returns \code{"Complete"}.
}
\details{
Call this function after \code{get_queryId_status()} returns \code{"Complete"}.
}
\examples{
\dontrun{
## Get a gzipped .sdf file
query_id <- "a valid 36-character ChemSpider apikey"
apikey <- "a valid 32-character ChemSpider apikey"
get_query_id_results_sdf(query_id = query_id, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/filter/{queryId}/results/sdf}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
