% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-get_queryId_results_sdf.R
\name{get_queryId_results_sdf}
\alias{get_queryId_results_sdf}
\title{GET the gzipped .sdf file for a ChemSpider query}
\usage{
get_queryId_results_sdf(queryId, status, apikey)
}
\arguments{
\item{queryId}{A valid 36-character ChemSpider \code{queryId}.}

\item{status}{A character string indicating the query status as returned by \code{chemspiderapi::get_queryId_status()}.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
Returns a (base64-encoded) character vector
}
\description{
This function is used to download a single .sdf file from ChemSpider after \code{chemspiderapi::get_queryId_status()} returns \code{"Complete"}.
}
\details{
Call this function after \code{chemspiderapi::get_queryId_status()} returns \code{"Complete"}.\cr
\cr
If successful, returns a character vector of results from the query, a gzipped .sdf file.\cr
\cr
If not successful, returns \code{NA}.\cr
\cr
This function is fully \code{tidyverse} compatible, e.g., for use in \code{purrr::map_chr()}.
}
\examples{
\dontrun{
## GET a gzipped .sdf file
queryId <- "a valid 36-character ChemSpider apikey"
apikey <- "a valid 32-character ChemSpider apikey"
get_queryId_results_sdf(queryId = queryId, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/filter/{queryId}/results/sdf}
}
