% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECORDS-get_record_id_details.R
\name{get_record_id_details}
\alias{get_record_id_details}
\title{Get details for a ChemSpider record}
\usage{
get_record_id_details(record_id, fields = "all", apikey = NULL)
}
\arguments{
\item{record_id}{A valid (integer) ChemSpider ID.}

\item{fields}{Either a single character string, a character vector, or a character list stating which fields to return. Alternatively, \code{"all"} returns all possible \code{fields}. \code{fields} is NOT case sensitive, but see details for a list of possible entries.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
A list containing the specified fields.
}
\description{
This function returns record details from ChemSpider.
}
\details{
"Call this endpoint with a Record ID as an integer.\cr
\cr
The available fields are: SMILES, Formula, InChI, InChIKey, StdInChI, StdInChIKey, AverageMass, MolecularWeight, MonoisotopicMass, NominalMass, CommonName, ReferenceCount, DataSourceCount, PubMedCount, RSCCount, Mol2D, Mol3D."
}
\examples{
\dontrun{
## Get the record details for caffeine
record_id <- 2424L
apikey <- "A valid 32-character Chemspider API key"
get_record_id_details(record_id = record_id, 
                      fields = c("SMILES", "Formula", "MolecularWeight", 
                                 "CommonName"), 
                      apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/records/{recordId}/details}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
