% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECORDS-get_record_id_external_references.R
\name{get_record_id_external_references}
\alias{get_record_id_external_references}
\title{Get external references of a record ID}
\usage{
get_record_id_external_references(
  record_id,
  data_sources = NULL,
  apikey = NULL
)
}
\arguments{
\item{record_id}{A valid (integer) ChemSpider ID.}

\item{data_sources}{Either a character string, a vector of characters, or a list of characters detailing which \code{data_sources} to look up. If left as is, will return all \code{data_sources}.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
A list containing the external references.
}
\description{
GET a list of external references for a ChemSpider ID.
}
\details{
It is recommended to specify which \code{data_sources} to load, as some substances have a substantial amount of references (>10'000). Use \code{chemspiderapi::get_datasources()} for a complete list of all \code{data_sources}.
}
\examples{
\dontrun{
## Get the PubChem external reference for caffeine
record_id <- 2424L
data_sources <- "PubChem"
apikey <- "a valid 32-character ChemSpider apikey"
get_record_id_external_references(record_id = record_id, 
                                  data_sources = data_sources, 
                                  apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/records/{recordId}/externalreferences}
}
