% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECORDS-get_recordId_externalreferences.R
\name{get_recordId_externalreferences}
\alias{get_recordId_externalreferences}
\title{Get external references of a record ID}
\usage{
get_recordId_externalreferences(
  recordId,
  dataSources = NULL,
  apikey,
  coerce = FALSE
)
}
\arguments{
\item{recordId}{A valid (integer) ChemSpider ID.}

\item{dataSources}{Either a character string, a vector of characters, or a list of characters detailing which \code{dataSources} to look up. If left as is, will return all \code{dataSources}.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}

\item{coerce}{\code{logical}: should the list be coerced to a data.frame? Defaults to \code{FALSE}.}
}
\value{
Either a data frame (if no options are dropped) or a character/integer vector, depending on the external ID.
}
\description{
GET a list of external references for a ChemSpider ID.
}
\details{
It is recommended to specify which \code{dataSources} to load, as some substances have a substantial amount of references (>10'000). Use \code{chemspiderapi::get_datasources()} for a complete list of the >350 \code{dataSources}.
}
\examples{
\dontrun{
## Get the PubChem external reference for caffeine
recordId <- 2424L
dataSources <- "PubChem"
apikey <- "a valid 32-character ChemSpider apikey"
get_recordId_externalreferences(recordId = recordId, dataSources = dataSources, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/records/{recordId}/externalreferences}
}
