% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-post_intrinsicproperty.R
\name{post_intrinsicproperty}
\alias{post_intrinsicproperty}
\title{Post an intrinsic property}
\usage{
post_intrinsicproperty(property, formula, complexity = "any",
  isotopic = "any", mass, range, orderBy = "recordId",
  orderDirection = "ascending", apikey)
}
\arguments{
\item{property}{See Details.}

\item{formula}{See Details.}

\item{complexity}{See Details.}

\item{isotopic}{See Details.}

\item{mass}{See Details.}

\item{range}{See Details.}

\item{orderBy}{A character string indicating by which parameter the results should be ordered; see Details.}

\item{orderDirection}{A character string indicating in which direction the results should be ordered; see Details.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
Returns the queryId string as (named) character vector.
}
\description{
Functionality to post an intrinsic property to obtain a \code{queryId} for use in \code{chemspiderapi::get_queryId_status()} and \code{chemspiderapi::get_queryId_results()}.
}
\details{
Valid entries for property are: \code{formula}, \code{molecularWeight}, \code{nominalMass}, \code{averageMass}, or \code{monoisotopicMass}.\cr
\cr
Says ChemSpider:\cr
\cr
\emph{"Valid values for \code{complexity} are \code{any}, \code{single}, or \code{multiple} whereby a compound with a complexity of multiple has more than one disconnected system in it or a metal atom or ion.\cr
\cr
Valid values for \code{isotopic} are \code{any}, \code{labeled}, or \code{unlabeled}.\cr
\cr
You can use \code{orderBy} and \code{orderDirection} to specify the sort order for the results. Valid values for \code{orderBy} are \code{recordId}, \code{massDefect}, \code{molecularWeight}, \code{referenceCount}, \code{dataSourceCount}, \code{pubMedCount}, \code{rscCount}.\cr
\cr
Valid values for \code{orderDirection} are \code{ascending}, \code{descending}."}\cr
\cr
If successful, returns the \code{queryId} as character vector.
}
\examples{
\dontrun{
## Post an intrinsic property for caffeine to get a query ID
property = "Formula"
formula = "C8H10N4O2"
apikey <- "a valid 32-character ChemSpider apikey"
post_intrinsicproperty(property = property, formula = formula, apikey = apikey)

property <- "MolecularWeight"
mass <- 194
range <- 0.5
apikey <- "a valid 32-character ChemSpider apikey"
post_intrinsicproperty(property = property, mass = mass, range = range, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/filter/intrinsicproperty}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
