% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONALITY-remove_apikey.R
\name{remove_apikey}
\alias{remove_apikey}
\title{Remove a ChemSpider API Key}
\usage{
remove_apikey(service = "ChemSpider API Key", username = NULL)
}
\arguments{
\item{service}{The name of the key ring service. Defaults to "ChemSpider API Key".}

\item{username}{The username. Defaults to the username as returned by \code{base::Sys.info()}.}
}
\value{
Securely removes the ChemSpider API key.
}
\description{
Functionality to remove a previously stored ChemSpider API key.
}
\details{
This function is useful to remove an existing ChemSpider API key, for example because of a typo in the API key.
}
\examples{
\dontrun{
## get an API key
remove_apikey()
}
}
