% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONALITY-remove_rate_limit.R
\name{remove_rate_limit}
\alias{remove_rate_limit}
\title{Remove a Rate Limit for chemspiderapi Functions}
\usage{
remove_rate_limit()
}
\value{
Removes any rate-limition of chemspiderapi functions.
}
\description{
Functionality to remove all rate limits for chemspiderapi functions interacting with ChemSpider's API services.
}
\details{
This function is useful to remove any rate limitations from chemspiderapi functionalities, as provided by \code{chemspiderapi::set_rate_limit()}.
}
\examples{
\dontrun{
## Set a rate limit for all chemspiderapi functions
rate <- 15
limit <- 60
set_rate_limit(rate = rate, limit = limit)

## Remove the rate limit
remove_rate_limit()
}
}
