% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONALITY-save_apikey.R
\name{save_apikey}
\alias{save_apikey}
\title{Save a ChemSpider API Key}
\usage{
save_apikey(service = "ChemSpider API Key", username = NULL)
}
\arguments{
\item{service}{The name of the key ring service. Defaults to "ChemSpider API Key".}

\item{username}{The username. Defaults to the username as returned by \code{base::Sys.info()}.}
}
\value{
Stores the ChemSpider API key as system key.
}
\description{
Functionality to interactively set a ChemSpider API key via keyring.
}
\details{
This function is useful to effectively and securely store your ChemSpider API key. The API is entered interactively in a pop-up console.\cr
\cr
To retrieve the ChemSpider API key, the function \code{chemspiderapi::get_apikey()} is provided.
}
\examples{
\dontrun{
## save an API key
save_apikey()
}
}
