% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONALITY-set_rate_limit.R
\name{set_rate_limit}
\alias{set_rate_limit}
\title{Set a Rate Limit for chemspiderapi Functions}
\usage{
set_rate_limit(rate, limit)
}
\arguments{
\item{rate}{Number (integer) of calls to make within a specific time limit.}

\item{limit}{Time range (integer) over which to apply the limit, in seconds.}
}
\value{
Creates rate-limited functions of all chemspiderapi functions interacting with ChemSpider's API services.
}
\description{
Functionality to set a rate limit for all chemspiderapi functions interacting with ChemSpider's API services.
}
\details{
This function is useful if you have rate limitations from ChemSpider.\cr
\cr
To "reset" the functions to a non-rate limited function, use \code{chemspiderapi::remove_rate_limit()}.
}
\examples{
\dontrun{
## Set a rate limit for all chemspiderapi functions
rate <- 15
limit <- 60
set_rate_limit(rate = rate, limit = limit)
}
}
