% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONALITY-update_rate_limit.R
\name{update_rate_limit}
\alias{update_rate_limit}
\title{Update a Rate Limit for chemspiderapi Functions}
\usage{
update_rate_limit(rate, limit)
}
\arguments{
\item{rate}{Number (integer) of calls to make within a specific time limit.}

\item{limit}{Time range (integer) over which to apply the limit, in seconds.}
}
\value{
Creates updated rate-limited functions of all chemspiderapi functions interacting with ChemSpider's API functions.
}
\description{
Functionality to update a rate limit for all chemspiderapi functions interacting with ChemSpider's API services on the fly.
}
\details{
This function is useful if you have to update rate limitations from ChemSpider.\cr
\cr
To "reset" the functions to a non-rate limited function, use \code{chemspiderapi::remove_rate_limit()}.
}
\examples{
\dontrun{
## Set a rate limit for all chemspiderapi functions
rate1 <- 15
limit1 <- 60
set_rate_limit(rate = rate1, limit = limit1)

## Update a rate limit for all chemspiderapi functions
rate2 <- 500
limit2 <- 86400
update_rate_limit(rate = rate2, limit = limit2)
}
}
