% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOKUPS-get_data_sources.R
\name{get_data_sources}
\alias{get_data_sources}
\title{Get external ChemSpider data sources}
\usage{
get_data_sources(apikey = NULL)
}
\arguments{
\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
A list of all external data sources.
}
\description{
Returns all external ChemSpider data sources.
}
\details{
"Many other endpoints let you restrict which sources are used to lookup the requested query. Restricting the sources makes queries faster."\cr
\cr
Returns all available external ChemSpider data sources.\cr
\cr
This function is most useful for narrowing down \code{data_sources} in other chemspiderapi functions, for example \code{get_record_id_external_references()}.
}
\examples{
\dontrun{
## Get external data sources of ChemSpider
apikey <- "A valid API key"
get_data_sources(apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/lookups/datasources}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
