% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOKUPS-get_datasources.R
\name{get_datasources}
\alias{get_datasources}
\title{Get external ChemSpider data sources}
\usage{
get_datasources(apikey)
}
\arguments{
\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
A character vector, length > 350.
}
\description{
Returns a vector of all external ChemSpider data sources.
}
\details{
"Many other endpoints let you restrict which sources are used to lookup the requested query. Restricting the sources makes queries faster."\cr
\cr
Returns a character vector with all available external ChemSpider data sources.\cr
\cr
This function is most useful for narrowing down \code{dataSources} in other chemspiderapi functions, e.g., \code{chemspiderapi::get_recordId_externalreferences()}.
}
\examples{
\dontrun{
## Get external data sources of ChemSpider
apikey <- "A valid 32-character Chemspider API key"
get_datasources(apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/lookups/datasources}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
