% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOKUPS.R
\name{get_datasources}
\alias{get_datasources}
\title{GET a listing of all external ChemSpider data sources}
\usage{
get_datasources(apikey)
}
\arguments{
\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
A unnamed character vector, length > 350
}
\description{
Returns a listing of all external ChemSpider data sources.
}
\details{
If succesfull, return a character vector listing all available (external) data sources.\cr
\cr
If not succesfull, returns a warning and \code{NA}.\cr
\cr
This function is most useful for narrowing down \code{dataSources} in other \code{chemspiderapi} functions, e.g., \code{chemspiderapi::get_references()}.
}
\examples{
\dontrun{
## GET the external data sources of ChemSpider
apikey <- "A valid 32-character Chemspider API key"
get_datasources(apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/lookups/datasources}
}
