% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOKUPS-get_datasources.R
\name{get_datasources}
\alias{get_datasources}
\title{Get External ChemSpider Data Sources}
\usage{
get_datasources(apikey, ...)
}
\arguments{
\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}

\item{...}{Additional parameters; currently not implemented.}
}
\value{
A unnamed character vector, length > 350
}
\description{
Returns a vectorized list of all external ChemSpider data sources.
}
\details{
If succesfull, returns a character vector listing all available (external) ChemSpider data sources.\cr
\cr
If not succesfull, returns an error.\cr
\cr
This function is most useful for narrowing down \code{dataSources} in other chemspiderapi functions, e.g., \code{chemspiderapi::get_references()}.
}
\examples{
\dontrun{
## GET the external data sources of ChemSpider
apikey <- "A valid 32-character Chemspider API key"
get_datasources(apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/lookups/datasources}
}
