% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-get_formula_batch_queryId_results.R
\name{get_formula_batch_queryId_results}
\alias{get_formula_batch_queryId_results}
\title{Get results of a formula batch query from ChemSpider}
\usage{
get_formula_batch_queryId_results(queryId, status, apikey, coerce = FALSE)
}
\arguments{
\item{queryId}{A valid 36-character query ID, as returned by \code{post_formula_batch()}.}

\item{status}{A character string indicating the query status as returned by \code{get_formula_batch_queryId_status()}.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}

\item{coerce}{\code{logical}: should the list be coerced to a \code{data.frame}? Defaults to \code{FALSE}.}
}
\value{
Returns the (integer) ChemSpider IDs as \code{list} or \code{data.frame}.
}
\description{
Get results of a formula batch query from ChemSpider after \code{get_formula_batch_queryId_status()} returns \code{"Complete"}.
}
\details{
Before running \code{get_formula_batch_queryId_results()}, make sure \code{get_formula_batch_queryId_status()} returns \code{"Complete"}.
}
\examples{
\dontrun{
## Obtain the result from a mass batch query
apikey <- "a valid 32-character ChemSpider apikey"
queryId <- "a valid 36-character ChemSpider queryId"
status <- get_formula_batch_queryId_status(queryId = queryId, apikey = apikey)
get_formula_batch_queryId_results(queryId = queryId, status = status, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/filter/formula/batch/{queryId}/results}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
