% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-get_mass_batch_queryId_results.R
\name{get_mass_batch_queryId_results}
\alias{get_mass_batch_queryId_results}
\title{Get results of a mass batch query from ChemSpider}
\usage{
get_mass_batch_queryId_results(queryId, status, apikey)
}
\arguments{
\item{queryId}{A valid 36-character query ID, as returned by \code{chemspiderapi::post_mass_batch()}.}

\item{status}{status A character string indicating the query status as returned by \code{chemspiderapi::get_mass_batch_queryId_status()}}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
Returns the record IDs of a mass batch query
}
\description{
Get results from a ChemSpider query after \code{chemspiderapi::get_mass_batch_queryId_status()} returns \code{"Complete"}.
}
\details{
Before running \code{chemspiderapi::get_mass_batch_queryId_results()}, make sure \code{chemspiderapi::get_mass_batch_queryIdstatus()} returns \code{"Complete"}.
}
\examples{
\dontrun{
## Obtain the result from a mass batch query
apikey <- "a valid 32-character ChemSpider apikey"
queryId <- "a valid 36-character ChemSpider queryId"
get_mass_batch_queryId_results(queryId = queryId, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/filter/mass/batch/{queryId}/results}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
