% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECORDS-get_recordId_details.R
\name{get_recordId_details}
\alias{get_recordId_details}
\title{GET the record details for a compound from ChemSpider}
\usage{
get_recordId_details(recordId, fields = "all", apikey, id = TRUE)
}
\arguments{
\item{recordId}{A valid (integer) ChemSpider ID.}

\item{fields}{Either a single character string, a character vector, or a character list stating which fields to return. Alternatively, \code{"all"} returns all possible \code{fields}. \code{fields} is NOT case sensitive, but see details for a list of possible entries.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}

\item{id}{\code{logical}: Should the \code{id} column (i.e., the \code{recordId}) be part of the output?}
}
\value{
A \code{data.frame} if multiple columns are returned, or a (named) vector of the appropriate type if only one \code{field} is returned.
}
\description{
This function is used to return record details from ChemSpider.
}
\details{
The available options for \code{fields} are: \code{"SMILES"}, \code{"Formula"}, \code{"AverageMass"}, \code{"MolecularWeight"}, \code{"MonoisotopicMass"}, \code{"NominalMass"}, \code{"CommonName"}, \code{"ReferenceCount"}, \code{"DataSourceCount"}, \code{"PubMedCount"}, \code{"RSCCount"}, \code{"Mol2D"}, \code{"Mol3D"}, and \code{"all"}.\cr
\cr
If successful, returns either a \code{data.frame} of results from the query with the requested fields; if only one paraemeter is returned, this is then transformed into a vector.\cr
\cr
If not successful, it returns \code{NA}.\cr
\cr
This function is fully \code{tidyverse} compatible, e.g., for use in \code{purrr::map_*()}.
}
\examples{
\dontrun{
## GET the record details for aspirin
recordId <- 2157L
apikey <- "A valid 32-character Chemspider API key"
get_recordId_details(recordId = recordId, 
                     fields = c("SMILES", "Formula", "MolecularWeight", "CommonName"), 
                     apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/records/{recordId}/details}
}
