% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECORDS-get_recordId_externalreferences.R
\name{get_recordId_externalreferences}
\alias{get_recordId_externalreferences}
\title{GET external references of a record ID}
\usage{
get_recordId_externalreferences(recordId, dataSources = NULL, apikey,
  source = TRUE, sourceUrl = TRUE, externalUrl = TRUE)
}
\arguments{
\item{recordId}{A valid (integer) ChemSpider ID.}

\item{dataSources}{Either a character string, a vector of characters, or a list of characters detailing which \code{dataSources} to look up. If left as is, will return all \code{dataSources}.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}

\item{source}{\code{logical}: Should the source name be returned (ChemSpider default)?}

\item{sourceUrl}{\code{logical}: Should the source URL be returned (ChemSpider default)?}

\item{externalUrl}{\code{logical}: Should the external URL be returned (ChemSpider default)?}
}
\value{
Either a data frame (if no options are dropped) or a character/integer vector, depending on the external ID.
}
\description{
GET a list of external references for a ChemSpider ID.
}
\details{
It is recommended to specify which \code{dataSources} to load, as some substances have a substantial amount of references (>10'000). Use \code{chemspiderapi::get_datasources()} for a complete list of the >300 \code{dataSources}.\cr
\cr
If successful, returns a data frame with four columns: \code{source}, \code{sourceUrl}, \code{externalId}, and \code{externalUrl}; unless any of them are removed by setting them to \code{FALSE} in the function call. In case only one parameter, e.g. the external ID, is returned, the result is a vector.\cr
\cr
If not successful, returns \code{NA}.\cr
\cr
This function is fully \code{tidyverse} compatible, e.g., for use in \code{purrr::map()}.
}
\examples{
\dontrun{
## GET the PubChem external reference for aspirin
recordId <- 2157L
dataSources <- "PubChem"
apikey <- "a valid 32-character ChemSpider apikey"
get_recordId_externalreferences(recordId = recordId, dataSources = dataSources, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/records/{recordId}/externalreferences}
}
