% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECORDS-get_recordId_image.R
\name{get_recordId_image}
\alias{get_recordId_image}
\title{GET a PNG image of a ChemSpider record ID}
\usage{
get_recordId_image(recordId, apikey)
}
\arguments{
\item{recordId}{A valid (integer) ChemSpider ID.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
A raw array.
}
\description{
This function is used to obtain a .png image file of a ChemSpider record ID, e.g., after \code{chemspiderapi::get_queryID_results()}.
}
\details{
If succesfull, returns a numeric (double) array.
}
\examples{
\dontrun{
## GET the .png image for aspirin
recordId <- 2157L
apikey <- "a_valid_ChemSpider_API_key"
get_recordId_image(recordId = recordId, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/records/{recordId}/image}
}
