% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECORDS-get_recordId_mol.R
\name{get_recordId_mol}
\alias{get_recordId_mol}
\title{Get the MOL file of a ChemSpider record}
\usage{
get_recordId_mol(recordId, apikey)
}
\arguments{
\item{recordId}{A valid (integer) ChemSpider ID.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
A character string containing the (human-readable) .MOL file.
}
\description{
This function is used to download a single .MOL file from ChemSpider.
}
\details{
"If successful, returns a SDF file containing the single record for the matching record."\cr
\cr 
Call this endpoint with \code{recordId} as an integer. To save the MOL/SDF file, see the vignette "Saving MOL Files of Chemicals".
}
\examples{
\dontrun{
## Get the MOL file for caffeine
recordId <- 2424L
apikey <- "a valid 32-character ChemSpider apikey"
get_recordId_mol(recordId = recordId, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/get/records/{recordId}/mol}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
