% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECORDS.R
\name{post_batch}
\alias{post_batch}
\title{POST a batch of ChemSpider IDs}
\usage{
post_batch(recordIds, fields = "all", apikey, id = TRUE)
}
\arguments{
\item{recordIds}{A vector of integer ChemSpider IDs.}

\item{fields}{A character string indicating which fields to return; see Details.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}

\item{id}{\code{logical}: Should the record ID be returned (ChemSpider default)?}
}
\value{
A data frame (if multiple fields are returned), or a vector of adequate type if only one field is requiered.
}
\description{
POST a batch query of up to 100 ChemSpider record IDs.
}
\details{
The available options for \code{fields} are: \code{"SMILES"}, \code{"Formula"}, \code{"AverageMass"}, \code{"MolecularWeight"}, \code{"MonoisotopicMass"}, \code{"NominalMass"}, \code{"CommonName"}, \code{"ReferenceCount"}, \code{"DataSourceCount"}, \code{"PubMedCount"}, \code{"RSCCount"}, \code{"Mol2D"}, \code{"Mol3D"}, and \code{"all"} (default).\cr
\cr
If a \code{recordId} is not found, it is silently dropped.\cr
\cr
If successful, returns a data frame with each row being a \code{recordId} and each column being one of \code{fields}.\cr
\cr
If not successful, returns \code{NA}.\cr
\cr
This function is fully \code{tidyverse} compatible, e.g., for use in \code{purrr::map()}.
}
\examples{
\dontrun{
## POST a query for the first 50 ChemSpider entries.
recordIds <- 1:50
apikey <- "a valid 32-character ChemSpider apikey"
post_batch(recordIds = recordIds, apikey = apikey, id = FALSE)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/records/batch}
}
