% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOOLS.R
\name{post_convert}
\alias{post_convert}
\title{Convert chemical identifiers}
\usage{
post_convert(input, inputFormat, outputFormat, apikey)
}
\arguments{
\item{input}{A character string in the specified \code{inputFormat}.}

\item{inputFormat}{A character string indicating which format the input has. Can be one of the following: \code{InChI}, \code{InChIKey}, \code{Mol}, or \code{SMILES}. See Details for possible conversions.}

\item{outputFormat}{A character string indicating which type of output is desired. Can be one of the following: \code{InChI}, \code{InChIKey}, \code{Mol}, or \code{SMILES}. See Details for possible conversions.}

\item{apikey}{A valid 32-character string with a valid key for ChemSpider's API services.}
}
\value{
A character string with the desired converted identifier
}
\description{
Functionality to convert between different chemical identifier formats: InChI, InChIKey, Mol, and SMILES.
}
\details{
Submit an identifier (SMILES, InChI, InChIKey or Mol) as a string and return an identifier in another format (SMILES, InChI, InChIKey or Mol).\cr
\cr
Allowed conversions:\cr
\cr
from \code{InChI} to \code{InChIKey}\cr
\cr
from \code{InChI} to \code{Mol} file\cr
\cr
from \code{InChI} to \code{SMILES}\cr
\cr
from \code{InChIKey} to \code{InChI}\cr
\cr
from \code{InChIKey} to \code{Mol} file\cr
\cr
from \code{Mol} file to \code{InChI}\cr
\cr
from \code{Mol} file to \code{InChIKey}\cr
\cr
from \code{SMILES} to \code{InChI}\cr
\cr
Note: The identifier names are NOT case sensitive!\cr
\cr
If successful, performs the desired conversion and stores the result as a named ("output") character string.\cr
\cr
If not successful, it returns \code{NA}.\cr
\cr
This function is fully \code{tidyverse} compatible, e.g., for use in \code{purrr::map_chr()}
}
\examples{
\dontrun{
## Convert the InChI string of aspirin to a SMILES formula
input <- "InChI=1S/C9H8O4/c1-6(10)13-8-5-3-2-4-7(8)9(11)12/h2-5H,1H3,(H,11,12)"
inputFormat <- "InChI"
outputFormat <- "SMILES"
apikey <- "A valid 32-character Chemspider API key"
post_convert(input = input, inputFormat = inputFormat, outputFormat = outputFormat,
             apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/tools/convert}
}
