% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOOLS-post_convert.R
\name{post_convert}
\alias{post_convert}
\title{Convert between chemical identifiers}
\usage{
post_convert(
  input,
  inputFormat,
  outputFormat,
  apikey,
  coerce = FALSE,
  simplify = FALSE
)
}
\arguments{
\item{input}{A character string in the specified \code{inputFormat}.}

\item{inputFormat}{A character string indicating which format the input has. Can be one of the following: \code{InChI}, \code{InChIKey}, \code{Mol}, or \code{SMILES}. See Details for possible conversions.}

\item{outputFormat}{A character string indicating which type of output is desired. Can be one of the following: \code{InChI}, \code{InChIKey}, \code{Mol}, or \code{SMILES}. See Details for possible conversions.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}

\item{coerce}{\code{logical}: should the list be coerced to a data.frame? Defaults to \code{FALSE}.}

\item{simplify}{\code{logical}: should the results be simplified to a vector? Defaults to \code{FALSE}.}
}
\value{
A character string with the desired converted identifier
}
\description{
Functionality to convert between different chemical identifier formats: InChI, InChIKey, Mol, and SMILES.
}
\details{
"Specify the input format as a string called 'inputFormat', and the output as a string called 'outputFormat'. Allowed conversions: from InChI to InChI Key, from InChI to Mol file, from InChI to SMILES, from InChIKey to InChI, from InChI Key to Mol file, from Mol file to InChI, from Mol file to InChI Key, from SMILES to InChI."\cr
\cr
The identifier names are NOT case sensitive!
}
\examples{
\dontrun{
## Convert the InChI string of caffeine to a SMILES formula
input <- "InChI=1S/C8H10N4O2/c1-10-4-9-6-5(10)7(13)12(3)8(14)11(6)2/h4H,1-3H3"
inputFormat <- "InChI"
outputFormat <- "SMILES"
apikey <- "A valid 32-character Chemspider API key"
post_convert(input = input, inputFormat = inputFormat, outputFormat = outputFormat, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/tools/convert}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
