% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOOLS-post_convert.R
\name{post_convert}
\alias{post_convert}
\title{Convert between chemical identifiers}
\usage{
post_convert(input, input_format, output_format, apikey = NULL)
}
\arguments{
\item{input}{A character string in the specified \code{input_format}.}

\item{input_format}{A character string indicating which format the input has. Can be one of the following: \code{InChI}, \code{InChIKey}, \code{Mol}, or \code{SMILES}. See Details for possible conversions.}

\item{output_format}{A character string indicating which type of output is desired. Can be one of the following: \code{InChI}, \code{InChIKey}, \code{Mol}, or \code{SMILES}. See Details for possible conversions.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
A list with one character string with the converted identifier
}
\description{
Functionality to convert between different chemical identifier formats: InChI, InChIKey, Mol, and SMILES.
}
\details{
"Specify the input format as a string called 'input_format', and the output as a string called 'output_format'. Allowed conversions: from InChI to InChIKey, from InChI to Mol file, from InChI to SMILES, from InChIKey to InChI, from InChIKey to Mol file, from Mol file to InChI, from Mol file to InChIKey, from SMILES to InChI."\cr
\cr
The identifier names are NOT case sensitive!
}
\examples{
\dontrun{
## Convert the InChI string of caffeine to a SMILES formula
input <- "InChI=1S/C8H10N4O2/c1-10-4-9-6-5(10)7(13)12(3)8(14)11(6)2/h4H,1-3H3"
input_format <- "InChI"
output_format <- "SMILES"
post_convert(
  input = input, 
  input_format = input_format, 
  output_format = output_format, 
  apikey
  )
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/tools/convert}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
