% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-post_element.R
\name{post_element}
\alias{post_element}
\title{Post elements to obtain a query ID}
\usage{
post_element(
  includeElements,
  excludeElements,
  includeAll = FALSE,
  complexity = "any",
  isotopic = "any",
  orderBy = "recordId",
  orderDirection = "ascending",
  apikey
)
}
\arguments{
\item{includeElements}{A character vector of elements to include; maximum length 15.}

\item{excludeElements}{A character vector of elements to exclude; maximum length 100.}

\item{includeAll}{\code{logical}: Only look for records containing ALL elements of \code{includeElement}?}

\item{complexity}{See Details.}

\item{isotopic}{See Details.}

\item{orderBy}{A character vector indicating by which parameter to order. Defaults to \code{recordId}; see Details for options.}

\item{orderDirection}{A character vector indicating in which direction to order; either \code{ascending} (default) or \code{descending}.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
Returns the queryId string as (named) character vector.
}
\description{
Functionality to post up to 15 elements to include and up to 100 elements to exclude to obtain a \code{queryId} for use in \code{chemspiderapi::get_queryId_status()} and \code{chemspiderapi::get_queryId_results()}.
}
\details{
Says ChemSpider:\cr
\cr
\emph{"Optionally, you can also submit \code{orderBy} and \code{orderDirection} to specify the sort order for the results. If you do not specify a value for \code{orderBy}, results are sorted by [\code{recordId}] by default."}\cr
\cr
Valid values for \code{orderBy} are \code{recordId}, \code{massDefect}, \code{molecularWeight}, \code{referenceCount}, \code{dataSourceCount}, \code{pubMedCount}, \code{rscCount}.\cr
\cr
Valid values for \code{orderDirection} are \code{ascending}, \code{descending}.\cr
\cr
The default behavior is to consider records which contain any of the elements listed in \code{includeElements}. Setting \code{includeAll} to \code{TRUE} will only consider records which contain all the elements listed in the \code{includeElements}.\cr
\cr
Valid values for \code{complexity} are \code{"any"}, \code{"single"}, or \code{"multiple"} whereby a compound with a complexity of multiple has more than one disconnected system in it or a metal atom or ion.\cr
\cr
Valid values for \code{isotopic} are \code{"any"}, \code{"labeled"}, or \code{"unlabeled"}."
}
\examples{
\dontrun{
## Post the elements for caffeine and exclude certain elements
includeElements <- c("C", "H", "N", "O")
excludeElements <- c("K", "Na", "Fe", "Cl", "F")
apikey <- "a valid 32-character ChemSpider apikey"
post_element(includeElements = includeElements, excludeElements = excludeElements, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/filter/element}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
