% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-post_formula_batch.R
\name{post_formula_batch}
\alias{post_formula_batch}
\title{Post a batch of chemical formulas}
\usage{
post_formula_batch(
  formulas,
  dataSources = NULL,
  orderBy = "recordId",
  orderDirection = "ascending",
  apikey,
  coerce = FALSE
)
}
\arguments{
\item{formulas}{A character vector of up to 100 chemical formulas.}

\item{dataSources}{Optional: Either a single character string or a vector of character string specifying the data sources. A list of possible data sources can be obtained from \code{chemspiderapi::get_datasources()}.}

\item{orderBy}{A character string indicating by which parameter the results should be ordered; see Details.}

\item{orderDirection}{A character string indicating in which direction the results should be ordered; see Details.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}

\item{coerce}{\code{logical}: should the list be coerced to a data.frame? Defaults to \code{FALSE}.}
}
\value{
Returns the queryId string as (named) character vector.
}
\description{
Functionality to post a batch of up to 100 formulas to obtain a \code{queryId} for use in \code{chemspiderapi::get_formula_batch_queryId_status()} and \code{chemspiderapi::get_formula_batch_queryId_results()}.
}
\details{
Possible values for \code{orderBy} are: \code{"recordId"} (default), \code{"massDefect"}, \code{"molecularWeight"}, \code{"referenceCount"}, \code{"dataSourceCount"}, \code{"pubmedCount"}, and \code{"rscCount"}.\cr
\cr
Possible values for \code{orderDirection} are: \code{"ascending"} (default) and \code{"descending"}.\cr
\cr
\emph{"If dataSources is not specified, all known sources are searched. This will take longer."}
}
\examples{
\dontrun{
## Post the formulas of caffeine and cocaine to get a query ID
formulas <- c("C9H8O4", "C17H21NO4")
apikey <- "a valid 32-character ChemSpider apikey"
post_formula_batch(formulas = formulas, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/filter/formula/batch}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
