% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-post_inchi.R
\name{post_inchi}
\alias{post_inchi}
\title{POST an InChI string}
\usage{
post_inchi(inchi, apikey)
}
\arguments{
\item{inchi}{A valid InChI string; see Details.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
Returns the queryId string as (named) character vector.
}
\description{
Functionality to POST an InChI string to obtain a \code{queryId} for use in \code{chemspiderapi::get_queryID_status()} and \code{chemspiderapi::get_queryId_results()}.
}
\details{
The validity criteria for InChI strings are outlined here: \url{https://www.inchi-trust.org/technical-faq/#2.8}. If certain criteria are not met by the input \code{inchi}, \code{chemspiderapi::post_inchi()} returns a warning message and \code{NA}, and does not perform an API query.\cr
\cr
If successful, returns the \code{queryId} as character string.
}
\examples{
\dontrun{
## POST the InChI string of Aspirin to get a queryId
inchi <- "InChI=1S/C9H8O4/c1-6(10)13-8-5-3-2-4-7(8)9(11)12/h2-5H,1H3,(H,11,12)"
apikey <- "a valid 32-character ChemSpider apikey"
post_inchi(inchi = inchi, apikey = apikey)}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/filter/inchi}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
