% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-post_inchi.R
\name{post_inchi}
\alias{post_inchi}
\title{POST an InChI string}
\usage{
post_inchi(inchi, apikey, coerce = FALSE, simplify = FALSE)
}
\arguments{
\item{inchi}{A valid InChI string; see Details.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}

\item{coerce}{\code{logical}: should the list be coerced to a data.frame? Defaults to \code{FALSE}.}

\item{simplify}{\code{logical}: should the results be simplified to a vector? Defaults to \code{FALSE}.}
}
\value{
Returns the queryId string as (named) character vector.
}
\description{
Functionality to POST an InChI string to obtain a \code{queryId} for use in \code{chemspiderapi::get_queryID_status()} and \code{chemspiderapi::get_queryId_results()}.
}
\details{
The validity criteria for InChI strings are outlined here: \url{https://www.inchi-trust.org/technical-faq/#2.8}. If certain criteria are not met by the input \code{inchi}, \code{chemspiderapi::post_inchi()} returns an error message and does not perform an API query.
}
\examples{
\dontrun{
## Post the InChI string of caffeine to get a queryId
inchi <- "InChI=1S/C8H10N4O2/c1-10-4-9-6-5(10)7(13)12(3)8(14)11(6)2/h4H,1-3H3"
apikey <- "a valid 32-character ChemSpider apikey"
post_inchi(inchi = inchi, apikey = apikey)}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/filter/inchi}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
