% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-post_inchikey.R
\name{post_inchikey}
\alias{post_inchikey}
\title{Post an InChIKey as ChemSpider query}
\usage{
post_inchikey(inchikey, apikey)
}
\arguments{
\item{inchikey}{A valid 27-character InChIKey; see Details.}

\item{apikey}{A valid 32-character string with a valid key for ChemSpider's API services.}
}
\value{
Returns the queryId string as (named) character vector.
}
\description{
Functionality to post an InChIKey to obtain a \code{queryId} for use in \code{chemspiderapi::get_status()} and \code{chemspiderapi::get_results()}.
}
\details{
The validity criteria for InChIKeys are outlined here: \url{https://www.inchi-trust.org/technical-faq/#13.1}. If certain criteria are not met by the input \code{inchikey}, \code{chemspideR::post_inchikey()} returns an error message (and does not perform an API query). In the case of a non-standard \code{inchikey}, a warning is issued but the query will be performed.
}
\examples{
\dontrun{
## Post the InChIKey of caffeine to obtain a queryId
inchikey <- "RYYVLZVUVIJVGH-UHFFFAOYSA-N"
apikey <- "A valid 32-character Chemspider API key"
post_inchikey(inchikey = inchikey, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/filter/inchikey}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
