% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-post_name.R
\name{post_name}
\alias{post_name}
\title{Post a name to obtain a ChemSpider query ID}
\usage{
post_name(
  name,
  orderBy = "recordId",
  orderDirection = "ascending",
  apikey,
  coerce = FALSE,
  simplify = FALSE
)
}
\arguments{
\item{name}{A character string of the compound name.}

\item{orderBy}{A character string indicating by which parameter the results should be arranged (NOT case sensitive); see Details.}

\item{orderDirection}{A character string indicating which in which direction the results should be arranged (NOT case sensitive); see Details.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}

\item{coerce}{\code{logical}: should the list be coerced to a data.frame? Defaults to \code{FALSE}.}

\item{simplify}{\code{logical}: should the results be simplified to a vector? Defaults to \code{FALSE}.}
}
\value{
Returns the queryId string as (named) character vector.
}
\description{
Functionality to POST the name of a compound and obtain a ChemSpider query ID for subsequent use in \code{chemspiderapi::get_queryId_status()} and \code{chemspiderapi::get_queryId_results()}.
}
\details{
Allowed entries for \code{orderBy} are: \code{"recordId"} (default), \code{"massDefect"}, \code{"molecularWeight"}, \code{"referenceCount"}, \code{"dataSourceCount"}, \code{"pubMedCount"}, and \code{"rscCount"}.\cr
\cr
Allowed entries for \code{orderDirection} are: \code{"ascending"} (default) and \code{"descending"}.
}
\examples{
\dontrun{
## Post the name of caffeine to obtain a query ID
name <- "caffeine"
apikey <- "a valid 32-character ChemSpider apikey"
post_name(name = name, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/filter/name}
}
