% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING.R
\name{post_name}
\alias{post_name}
\title{POST a name to obtain a ChemSpider query ID}
\usage{
post_name(name, orderBy = "recordId", orderDirection = "ascending",
  apikey)
}
\arguments{
\item{name}{A character string of the compound name.}

\item{orderBy}{A character string indicating by which parameter the results should be arranged (NOT case sensitive); see Details.}

\item{orderDirection}{A character string indicating which in which direction the results should be arranged (NOT case sensitive); see Details.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
Returns the query ID as character vector.
}
\description{
Functionality to POST the name of a compound and obtain a ChemSpider query ID for subsequent use in \code{chemspiderapi::get_queryId_status()} and \code{chemspiderapi::get_queryId_results()}.
}
\details{
Allowed entries for \code{orderBy} are: \code{"recordId"} (default), \code{"massDefect"}, \code{"molecularWeight"}, \code{"referenceCount"}, \code{"dataSourceCount"}, \code{"pubMedCount"}, and \code{"rscCount"}.\cr
\cr
Allowed entries for \code{orderDirection} are: \code{"ascending"} (default) and \code{"descending"}.\cr
\cr
If successful, returns a 36-character \code{queryId} as a character vector. The \code{queryId} can be used in \code{chemspiderapi::get_queryId_status()} and \code{chemspider::get_queryId_results()}.\cr
\cr
If not successful, returns \code{NA}.\cr
\cr
This function is fully \code{tidyverse} compatible, e.g., for use in \code{purrr::map_chr()}.\cr
}
\examples{
\dontrun{
## POST the name of aspirin to obtain a query ID
name <- "aspirin"
apikey <- "a valid 32-character ChemSpider apikey"
post_name(name = name, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/filter/name}
}
