% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILTERING-post_smiles.R
\name{post_smiles}
\alias{post_smiles}
\title{POST a SMILES string}
\usage{
post_smiles(smiles, apikey)
}
\arguments{
\item{smiles}{A SMILES character string.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
Returns the queryId string as (named) character vector.
}
\description{
Functionality to POST a SMILES string to obtain a \code{queryId} for use in \code{chemspiderapi::get_queryID_status()} and \code{chemspiderapi::get_queryId_results()}.
}
\details{
If successful, returns the \code{queryId} as character string.\cr
}
\examples{
\dontrun{
## POST the SMILES string of aspirin to get a queryId
smiles <- "CC(=O)OC1=CC=CC=C1C(=O)O"
apikey <- "a valid 32-character ChemSpider apikey"
post_smiles(smiles = smiles, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/filter/smiles}
}
