% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOOLS-post_validate_inchikey.R
\name{post_validate_inchikey}
\alias{post_validate_inchikey}
\title{Validate an InChIKey}
\usage{
post_validate_inchikey(inchikey, apikey)
}
\arguments{
\item{inchikey}{A 27-character InChIKey to be validated.}

\item{apikey}{A 32-character string with a valid key for ChemSpider's API services.}
}
\value{
A logical indicating the validity of the POSTed InChIKey
}
\description{
Functionality to check the validity of an InChIKey.
}
\details{
If successful (i.e., the InChIKey is valid), returns \code{TRUE}.\cr
\cr
If not successful (i.e., the InChIKey is not valid), returns \code{FALSE}.\cr
\cr
Before this functions performs an API query, it runs quality checking as lined out at \url{https://www.inchi-trust.org/technical-faq/#13.1}. If an InChIKey is ruled out based on these criteria, it returns \code{FALSE} with a warning message.
}
\examples{
\dontrun{
## validate the InChIKey of aspirin
inchikey <- "BSYNRYMUTXBXSQ-UHFFFAOYSA-N"
apikey <- "a_valid_ChemSpider_API_key"
post_validate_inchikey(inchikey = inchikey, apikey = apikey)
}
}
\seealso{
\url{https://developer.rsc.org/compounds-v1/apis/post/tools/validate/inchikey}
}
