% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONALITY-set_rate_limit.R
\name{set_rate_limit}
\alias{set_rate_limit}
\title{Set a rate limit for chemspiderapi}
\usage{
set_rate_limit(rate, limit)
}
\arguments{
\item{rate}{(Integer) number of calls to make in a specific time limit.}

\item{limit}{(Double) Time range over which to apply the limit, in seconds.}
}
\value{
Creates rate-limited functions of (most) chemspiderapi functions
}
\description{
Functionality to set a rate limit for (nearly) all chemspiderapi functions
}
\details{
This function is useful if you have rate limitations from ChemSpider.\cr
\cr
To "reset" the functions to a non-rate limited function, use \code{chemspiderapi::remove_rate_limit()}.
}
\examples{
\dontrun{
## Set a rate limit for all chemspiderapi functions
rate <- 15
limit <- 60
set_rate_limit(rate = rate, limit = limit)
}
}
