% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRITING-write_image.R
\name{write_image}
\alias{write_image}
\title{Write a .png image file}
\usage{
write_image(image, file)
}
\arguments{
\item{image}{Either a base64-encoded character string or a numeric (double) array of the .png image.}

\item{file}{A character string indicating where and under which name to save the image.}
}
\value{
writes a .png image file
}
\description{
Convenience function to write a .png image file, e.g., after downloading with \code{chemspiderapi::get_recordId_image()}.
}
\details{
A convenience function to write downloaded .png images, as obtained from \code{chemspiderapi::get_recordId_image()}.\cr
\cr
If successful, it writes a 150 x 150 px .png image (as returned by ChemSpider).\cr
\cr
If not successful, returns \code{NA} without writing an image file.\cr
\cr
This function is fully \code{tidyverse} compatible, e.g., for use in \code{purrr::map()}.
}
\examples{
\dontrun{
## Saving the .png file for aspirin
recordId <- 2157L
apikey <- "a valid 32-character ChemSpider apikey"
image <- get_recordId_image(recordId = recordId, apikey = apikey)
write_image(image = image, file = "aspirin.png")
}
}
