% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRITING-write_mol.R
\name{write_mol}
\alias{write_mol}
\title{Write a .MOL file}
\usage{
write_mol(mol, file)
}
\arguments{
\item{mol}{A valid .MOL file, e.g., downloaded via \code{chemspiderapi::get_recordId_mol()}.}

\item{file}{A character string indicating where and under which name to save the .MOL file.}
}
\value{
writes a (human readable) .MOL file
}
\description{
Convenience function to write a .MOL file, e.g., after downloading with \code{chemspiderapi::get_recordId_mol()}.
}
\details{
A convenience function to write .MOL files.\cr
\cr
If successful, it writes a .mol file.\cr
}
\examples{
\dontrun{
## Saving the .MOL file for aspirin
recordId <- 2157L
apikey <- "a_valid_ChemSpider_API_key"
mol <- get_recordId_mol(recordId = recordId, apikey = apikey)
write_mol(mol = mol, file = "aspirin.mol")
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
